tests: Add coverage of nbd_aio_opt_* in wrong state
Enhance the regression tests to prove that the completion callback is
not reached if the aio call itself reports an error; only the .free
callback is guaranteed.

Also add some asserts to the library code that may aid future readers
in seeing how we track transfer semantics of a callback.

Goes hand in hand with the documentation cleanups made in the previous
patch, but done separately for ease of backporting as nbd_aio_opt
calls are not in as many releases.

Reported-by: Tage Johansson <tage.j.lists@posteo.net>
Signed-off-by: Eric Blake <eblake@redhat.com>
Message-ID: <20230713193138.1306093-3-eblake@redhat.com>
Acked-by: Laszlo Ersek <lersek@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>