internal: Track chunk payload length left
While working on later patches, I noticed that I was frequently
referring to the wire contents of h->sbuf.reply.hdr.structured.length,
byte-swapping it, then repeating open-coded math for how many bytes
were consumed in earlier states.  It's easier to normalize the
remaining bytes to be parsed into a variable that persists across the
life of the reply, and also makes it possible to track that by the end
of any structured reply, we have dealt with all of the bytes that the
header advertised.  Some of the worst implicit assumptions in block
status code will be fixed in the next patch with the help of this new
variable.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-ID: <20230721162411.3312606-2-eblake@redhat.com>
Acked-by: Laszlo Ersek <lersek@redhat.com>