block_status: Refactor array storage
For 32-bit block status, we were able to cheat and use an array with
an odd number of elements, with array[0] holding the context id, and
passing &array[1] to the user's callback.  But once we have 64-bit
extents, we can no longer abuse array element 0 like that, for two
reasons: 64-bit extents contain uint64_t which might not be
alignment-compatible with an array of uint32_t on all architectures,
and the new NBD_REPLY_TYPE_BLOCK_STATUS_EXT adds an additional count
field before the array.  What's more, there's no need to byte-swap the
entries if we won't be calling the user's callback, when the server
sent us an unknown context id.

Split out a new state CHUNK_REPLY.BS_HEADER to receive the context id
(and eventually the new count field for 64-bit replies) separately
from the extents array.  Add another type nbd_chunk_block_status_32 in
the payload section for tracking it (with a name anticipating the
64-bit counterpart type under extended headers).  Add a helper
function bs_reply_length_ok() to enable more assertions that our
memory management is sane without quite so much open-coding of magic
values.  With these changes, the byteswap of the entries can be
delayed until we know we need to call the user's callback.  No
behavioral change, other than the rare possibility of landing in the
new state.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-ID: <20230721162411.3312606-3-eblake@redhat.com>
Acked-by: Laszlo Ersek <lersek@redhat.com>