states: Prepare to receive 64-bit replies
Support receiving headers for 64-bit replies if extended headers were
negotiated.  We already insist that the server not send us too much
payload in one reply, so we can exploit that and continue to use
h->payload_left as a 32-bit length for the rest of the payload
parsing.  The NBD protocol specifically documents that extended mode
takes precedence over structured replies, and that there are no simple
replies in extended mode.  We can also take advantage that the cookie
field is in the same offset in all the various reply types.  Checking
that the server sent back the correct offset adds a bit more verbosity
for the sake of nicer debug messages.

Note that if we negotiate extended headers, but a non-compliant server
replies with a non-extended header, this patch will stall waiting for
the server to send more bytes rather than noticing that the magic
number is wrong (for aio operations, you'll get a magic number
mismatch once you send a second command that elicits a reply; but for
blocking operations, we basically deadlock).  The easy alternative
would be to read just the first 4 bytes of magic, then determine how
many more bytes to expect; but that would require more states and
syscalls, and not worth it since the typical server will be compliant.
The other alternative is what the next patch implements: teaching
REPLY.RECV_REPLY to handle short reads that were at least long enough
to transmit magic to specifically look for magic number mismatch.

At this point, h->extended_headers is permanently false (we can't
enable it until all other aspects of the protocol have likewise been
converted).

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-ID: <20230721162411.3312606-6-eblake@redhat.com>
Acked-by: Laszlo Ersek <lersek@redhat.com>