states: Break deadlock if server goofs on extended replies
One of the benefits of extended replies is that we can do a
fixed-length read for the entire header of every server reply, which
is fewer syscalls than the split-read approach required by structured
replies.  But one of the drawbacks of doing a large read is that if
the server is non-compliant (not a problem for normal servers, but
something I hit rather more than I'd like to admit while developing
extended header support in servers), nbd_pwrite() and friends will
deadlock if the server replies with the wrong header.  Add in some
code to catch that failure mode and move the state machine to DEAD
sooner, to make it easier to diagnose the fault in the server.

Unlike in the case of an unexpected simple reply from a structured
server (where, since b31e7bac, we can merely fail the command with
EPROTO and successfully move on to the next server reply, because we
didn't read too many bytes yet), in this case we really do have to
move to DEAD: we cannot assume our short read was just the 16 (simple)
or 20 (structured) bytes that the server sent for this command,
because it is also possible that we can see a short read even while
reading two back-to-back replies; if we have already read the initial
bytes of the server's next reply, we have no way to push those extra
bytes back onto our read stream for parsing on our next pass through
the state machine.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-ID: <20230721162411.3312606-7-eblake@redhat.com>
Acked-by: Laszlo Ersek <lersek@redhat.com>