golang: Export meta-context constants
Go insists that a module's members are not public unless they start
with a capital letter.  Our attempt at defining constants
'namespace_base' and 'context_base_allocation' were thus only visible
to our in-package tests, and not for external clients.  For now, we
stick with ALL_CAPS instead of MixedCase constants; as Go bindings do
not have the same API constraints as C, we could change that in the
future if needed.

Although this patch does not completely address gofmt complaints,
adding some strategic grouping comments inside the const() block has
an additional benefit of reducing the amount of whitespace that gofmt
wants to inject into unrelated lines.

The change to generated golang/bindigs.go is:

| --- golang/bindings.go.bak	2023-07-26 12:06:47.618324437 -0500
| +++ golang/bindings.go	2023-07-26 12:06:55.521385295 -0500
| @@ -111,14 +111,18 @@
|      READ_DATA uint32 = 1
|      READ_HOLE uint32 = 2
|      READ_ERROR uint32 = 3
| -    namespace_base = "base:"
| -    context_base_allocation = "base:allocation"
| +    /* Meta-context namespace "base" */
| +    NAMESPACE_BASE = "base:"
| +    CONTEXT_BASE_ALLOCATION = "base:allocation"
| +    /* Defined bits in "base:allocation" */
|      STATE_HOLE uint32 = 1
|      STATE_ZERO uint32 = 2
| -    namespace_qemu = "qemu:"
| -    context_qemu_dirty_bitmap = "qemu:dirty-bitmap:"
| +    /* Meta-context namespace "qemu" */
| +    NAMESPACE_QEMU = "qemu:"
| +    CONTEXT_QEMU_DIRTY_BITMAP = "qemu:dirty-bitmap:"
| +    /* Defined bits in "qemu:dirty-bitmap:" */
|      STATE_DIRTY uint32 = 1
| -    context_qemu_allocation_depth = "qemu:allocation-depth"
| +    CONTEXT_QEMU_ALLOCATION_DEPTH = "qemu:allocation-depth"
|  )
|
|  /* SetDebug: set or clear the debug flag */

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-ID: <f4196229-a2c9-37de-27d5-f74ff5da061b@redhat.com>
Acked-by: Laszlo Ersek <lersek@redhat.com>
[eblake: Mention API considerations]