golang: Tweak comment formatting in test 590
Running 'gofmt' on the pre-patch contents of test 590 misformats a
couple of multi-line comments: go assumes that when a multi-line
comment appears in a documentation context, with the second line being
indented with only whitespace, then the first is a one-line doc
comment summary and the second is a new paragraph with more details.
For one of our three comments, this is true, but for the other two,
gofmt's suggestion is bogus.  Fix that by using leading * to mark
comment continuations, at which point gofmt is happy with our file
as-is (although I did not actually try to generate documentation to
see if further tweaks might be needed to improve such rendering).

Signed-off-by: Eric Blake <eblake@redhat.com>