golang: Improve whitespace style in generated bindings
Improve our generated output to more closely align to the style
recommended by 'gofmt': mainly no space before opening '(' and better
use of blank lines.

However, I am _intentionally_ sticking with 4-space indent[1]
(consistent with our settings in .editorconfig) rather than with TAB,
as putting raw TABs in GoLang.ml is going to cause problems with our
OCaml formatting, and using \t escapes makes the code snippets harder
to read.  It is possible to prefilter things to limit gofmt output to
just non-TAB issues as follows:

  sed ':a;s/^    \( *\)/\1\t/;ta' < golang/bindings.go | gofmt -d

and similarly for closures.go and wrappers.go.  In fact, the only
remaining differences are that we do not produce columnar output in
lists (such as const() declarations where we don't line up '=') in
bindings.go; fixing that would require more OCaml work to prepare the
lists in two passes (once to learn width, the second to do padded
output).  And since we _aren't_ going to output TAB, it is just as
easy to state that we may someday incorporate an optional use of gofmt
directly into our build process, and that the official tarballs will
be properly formatted that way, but all other developers can get by
just fine with Go code that may not be formatted correctly but at
least compiles.

[1] Surprisingly, Go does NOT mention what tab stops it prefers[2],
which means the only sane approach is to avoid mid-line TABs.  'gofmt'
only ever uses leading TABs (if it were to use mid-line TABs, you
could use columnar output to reverse engineer what tab-stops it
prefers, but you would also be unable to use tweaks like .editorconfig
with a tabstop of 4 to not have indents take up so much screen space).

[2] https://go.dev/doc/effective_go#formatting

Signed-off-by: Eric Blake <eblake@redhat.com>