block_status: Add some sanity checking of server lengths
Previously, we had not been doing any validation of server extent
responses, which means a client query at an offset near the end of the
export can result in a buggy server sending a response longer than the
export length and potentially confusing the client.  The NBD spec also
says that an extent length should be non-zero so that a successful
block status call makes progress.  It is easy enough to track that the
server has not overflowed the export size, and that we ensure an error
on no progress even when the buggy server claims success.  Since the
spec says a client should be prepared for a block status result to be
truncated, the client should not care whether the truncation happened
at the server or at libnbd after validating the server's response.

In the process, this patch reorganizes some of the code so that early
exits are obvious, leading for less indentation in the success path.

Adding this sanity checking now makes it easier for future patches to
do orthogonal support for a server's 32- or 64-bit reply, vs. a
client's 32- or 64-bit API call.  Once 64-bit replies are in play, we
will additionally have to worry about a 64-bit reply that overflows a
32-bit API callback without exceeding the exportsize.  Similarly,
since nbd_get_size() already caps export size at 63 bits (based on
off_t limitations), we have guaranteed that a 64-bit API callback will
never see an extent length that could appear negative in a 64-bit
signed type (at least OCaml benefits from that guarantee, since its
only native 64-bit integer type is signed).

[Side note: nbdkit allows plugins to report 0-length extents for ease
of programming, but it sanitizes them so that the wire content never
sends a zero-length extent; actually testing how this code behaves in
the presence of a zero-length extent requires the creation of a
one-off specialized server]

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-ID: <20230803015720.1090260-2-eblake@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>