block_status: Accept 64-bit extents during block status
Support a server giving us a 64-bit extent.  Note that the protocol
says a server should not give a 64-bit answer when extended headers
are not negotiated; we can handle that by reporting EPROTO but
otherwise accepting the information.  Conversely, when extended
headers are in effect, the server must respond with the 64-bit extent
type, even if the extent length and flags both fit in 32 bits and the
client is using the 32-bit API (including the most-likely case where a
libnbd client was compiled against an older libnbd that lacked 64-bit
API, but is now running against a newer libnbd.so that negotiates
extended headers by default).  Note, however, that at the time of THIS
patch, there are still enough other things lacking that it is not yet
possible for extended headers to be enabled.

This patch updates the code to separate where the server data is read
(bs_raw) from what is provided to the API callback (bs_cooked, at this
point, still just 32-bit, although upcoming additions for the 64-bit
client API will make bs_cooked a union as well).  Ensure that both
arrays are allocated at the same point within the state machine (even
if we later end up not using bs_cooked for other reasons), so that we
aren't having to figure out how to deal with ENOMEM at other points in
time.  This puts twice as much memory pressure on a libnbd handle when
receiving a block status reply, but as we bound the server into
providing no more than 64M of payload (and the NBD spec recommends no
more than 1M extents), we are unlikely to see memory pressure failures
because of this change.

With this in place, the server's reply size and the client's API size
can be orthogonal; 32->64 and 64->64 will be implemented in later
patches adding new API, 32->32 continues to work as before, and we now
support 64->32, where we have to be careful of some narrowing
considerations:

- If the server replies with a 64-bit length but still within the
  exportsize, we don't want the request to fail at all, but we do have
  to narrow the length to fit into 32 bits.  This is okay, because the
  NBD protocol says a client must be prepared for a truncated result,
  and the client can't tell whether the server or libnbd did the
  truncation.  And this scenario is actually likely: because we are
  NOT planning on using symbol version aliasing, an app compiled
  against an older libnbd but run against a newer libnbd.so will
  automatically negotiate extended headers even though it can only use
  the 32-bit API.

- If the server replies with a 64-bit length larger than the
  exportsize, we want outright EPROTO failure.  We do not want to risk
  the server sending junk that could look like a negative number when
  treated as an int64_t instead of a uint64_t; and nbd_get_size() has
  already implicitly got us to the point that exportsize will never
  exceed the 63 bits of off_t.  Doing this also makes it easier to
  prove that our math involving 'total' does not overflow.  Note that
  on this path, we may set 'stop = true' more than once, but that's
  okay.

- If the server replies with flags larger than 32 bits, we can't
  represent the flag to the 32-bit callback.  If there were earlier
  extents, we want the callback to see those without an error (back to
  the idea of truncation being okay); but if this is the first extent,
  we have to report EOVERFLOW failure.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-ID: <20230803015720.1090260-9-eblake@redhat.com>