generator: Prepare for extent64 callback
Although we don't yet have an API that takes an extent64 callback,
it's easier to add the code here separately from the API.  We need to
temporarily mark a couple of helper functions in language bindings
with '__attribute__((unused))' or similar to keep the compiler from
warning about them.

The biggest change in this patch is that the array to be passed to the
callback is now populated based on which style API was in use, and
demonstrates that the API is orthogonal from the server's reply size.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-ID: <20230803015720.1090260-10-eblake@redhat.com>