golang: Optionally use gofmt on generated .go files
Combine my recent work on improving the generated Go output with
Tage's work on using a canonical formatter for Rust.  If gofmt is
available during the build, then the generated .go files will now use
TAB indents and have proper columnar alignment; if it is not
available, the project still compiles.

Signed-off-by: Eric Blake <eblake@redhat.com>