copy/copy-file-to-qcow2-compressed.sh: Skip test for all zeroes disk
Occasionally this test will choose a random seed which results in an
all-zeroes disk.  The test tries to convert this to a compressed qcow2
file, and fails because no compressed clusters are detected in the
resulting file.  This happens because qcow2 stores zero clusters with
a special sparse representation, they are never stored compressed, so
a disk with only zeroes in it will never contain compressed clusters.

To fix this, detect an all-zeroes disk and skip.

Reported-by: Eric Blake