rust: Allow rust compilation without rustfmt
Our goal is that rustfmt should be an optional dependency.  If
present, we use it to make generated .rs files look nicer in the
tarball created by 'make dist'.  But if absent, things should still
compile just fine.

However, the configure test currently rejects cargo as unusable if
'cargo fmt' doesn't work, which is overly strict.  Not to mention that
'test ... -a ...' is non-portable (we don't need two conditions, but
if we did, the portable spelling is 'test ... && test ...').

Fixes: b2511d640b ("rust: create basic Rust bindings", v1.17.3)
Signed-off-by: Eric Blake <eblake@redhat.com>