generator: Support Extent64 arg in C code
See the earlier commit "Add Extent64 arg type" for rationale in
supporting a new generator arg type.  This patch adds the C bindings
for use of Extent64: it exposes a new 'struct nbd_extent' in
<libnbd.h>, as well as the code to generate the C bindings for passing
an extent64_closure once a new API uses the type in an upcoming patch.

Note that 'struct nbd_block_descriptor_64' in lib/nbd-protocol.h is
exactly the same as what we want to use in C.  But it is easier to
stick a new public type in <libnbd.h> than to figure out how to expose
just part of a header we only want to use internally, and leaves us
free to alter nbd-protocol.h (which is shared with nbdkit) without
affecting public API.

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-ID: <20230803015720.1090260-4-eblake@redhat.com>
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>