golang: Utilize unsafe.Slice()
Now that the minimum Go version is new enough, we can simplify our
code that wants to convert a C array into a Go slice.

In particular, note that this was foreshadowed in commit 6725fa0e12,
where we were previously relying on the Go hack of accessing a C array
cast to an array of arbitrary size before then relying on any
optimizations in Go's copy() for a bulk transfer of the slice over to
a fresh Go object.  As before, I have not done any benchmark timings
to see if we are encountering any slight speed change by doing
element-wise copies instead of relying on copy(); but also as before,
I suspect the overhead of the copy to be in the noise compared to the
network overhead of getting the data in the first place.

This also fixes a theoretical latent bug spotted by Laszlo: namely, we
aren't guaranteed that 'C.size_t' is the same size as native 'int'; if
'int' is smaller, our cast to int during the call to make() could
truncate a large size and end up creating a smaller array than
intended.  Fortunately, we already know that our state machine will
not reach the callback for any server sending more than 64M of
extents, and therefore 'count' is smaller than 64M in the first place,
so dropping the cast does not have any observable effects, all while
avoiding the question about truncation in the first place.

Suggested-by: Laszlo Ersek <lersek@redhat.com>
Suggested-by: Nir Soffer <nsoffer@redhat.com>
Signed-off-by: Eric Blake <eblake@redhat.com>