ci: Skip nbd-server interop tests on Alpine
CI builds have been timing out on Alpine 3.18 and Alpine-edge (but not
3.17) with what seems to be a deadlock on interop/interop-nbd-server;
I'm not sure which version of nbd-server is installed (hence adding a
breadcrumb), but since we used to skip the test on older Ubuntu,
skipping it on Alpine for the short term seems reasonable.  Meanwhile,
prune out skipped entries that are no longer relevant after recent CI
updates.

Signed-off-by: Eric Blake <eblake@redhat.com>