rust: Check that compilation works with default features disabled
The Rust crate has some default features, (tokio and log), to enable the
asynchronous API and logging facilities. This patch extends the test
suite to check that the crate compiles even with these features
disabled. The crate did in fact not compile without those features,
so the patch adds a "#[allow(unused)]" attribute to the utils
module to mittigate an error.